// ui/src/lib/gen/OpenPaneRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PaneContent } from "./PaneContent";
import type { Placement } from "./Placement";
import type { WorkItemId } from "./WorkItemId";

export type OpenPaneRequest = { content: PaneContent, placement: Placement, focus: boolean, tab_title: string | null, work_item_id: WorkItemId | null, };