// ui/src/lib/gen/Tab.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LayoutNode } from "./LayoutNode";
import type { PaneId } from "./PaneId";
import type { TabId } from "./TabId";
import type { WorkItemId } from "./WorkItemId";

export type Tab = { id: TabId, title: string, work_item_id: WorkItemId | null, root: LayoutNode, focused_pane: PaneId | null, zoomed_pane: PaneId | null, };